refactor: move utils.ts to core directory
